+++ src/commcare_cloud/fab/operations/supervisor.py
@@ -7,12 +7,11 @@ import posixpath
 from contextlib import contextmanager
 import uuid
 
-from commcare_cloud.commands.celery_utils import get_celery_worker_name
 from commcare_cloud.commands.ansible.helpers import (
     get_django_webworker_name,
     get_formplayer_instance_name,
     get_formplayer_spring_instance_name,
-)
+    get_celery_worker_name)
 from fabric.api import roles, parallel, env, sudo, serial, execute
 from fabric.colors import magenta
 from fabric.context_managers import cd
